PGCL: enable fault-around, fix rmap/RSS accounting, fix remap_pte_range
Enable filemap_map_pages fault-around for PAGE_MMUSHIFT>0 by implementing
proper MMUPAGE-granularity PTE mapping in the fault-around path.

Key changes:

- filemap_set_ptes_cluster(): New function that maps PAGE_MMUCOUNT PTEs
  for a single kernel page during fault-around, with VMA bounds clamping
  to prevent leaking PTEs across VMA boundaries (e.g., mapping data VMA's
  NX PTEs into text VMA address space at ELF segment boundaries).

- Fix rmap/RSS accounting mismatch: zap_pte_range removes rmap refs
  per-PTE, so the add side must also add per-PTE. For small folios,
  folio_add_file_rmap_ptes() always does atomic_inc regardless of nr
  parameter, so call it once per PTE mapped.

- remap_pte_range(): Step by MMUPAGE_SIZE instead of PAGE_SIZE, track
  physical byte address, use pte_mksub() for sub-page offset within
  kernel pages.

- do_fault_around(): Convert units from PAGE to MMUPAGE for nr_ptes
  and vma_mmupages calculations.

- Widen estack_pages.size from u16 to u32 to fix PAGE_MMUSHIFT=3 build
  (EXCEPTION_STKSZ=65536 overflows u16).

Tested: PAGE_MMUSHIFT=0 and PAGE_MMUSHIFT=2 boot cleanly with all 8
comprehensive tests passing (file I/O, fork, large file, pipe, exec,
/proc/self/maps, /proc/self/smaps, brk/malloc). No bad page maps, no
RSS counter errors, no kernel BUGs.

+++ arch/x86/kernel/dumpstack_64.c
@@ -64,7 +64,7 @@ const char *stack_type_name(enum stack_type type)
  */
 struct estack_pages {
 	u32	offs;
-	u16	size;
+	u32	size;
 	u16	type;
 };
